fixing custom error decoder when response is not default
