+++ moove/infrastructure/src/main/kotlin/io/charlescd/moove/infrastructure/configuration/SimpleFeignEncoderConfiguration.kt
@@ -16,28 +16,16 @@
 
 package io.charlescd.moove.infrastructure.configuration
 
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties
-import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
 import feign.Logger
-import feign.Response
 import feign.codec.Encoder
 import feign.codec.ErrorDecoder
 import feign.form.FormEncoder
-import io.charlescd.moove.domain.MooveErrorCode
-import io.charlescd.moove.domain.exceptions.BusinessException
-import java.io.IOException
-import java.lang.Exception
-import java.lang.IllegalArgumentException
-import java.lang.RuntimeException
-import java.nio.charset.StandardCharsets
-import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.ObjectFactory
 import org.springframework.boot.autoconfigure.http.HttpMessageConverters
 import org.springframework.cloud.openfeign.support.SpringEncoder
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.context.annotation.Scope
-import org.springframework.util.StreamUtils
 
 @Configuration
 class SimpleFeignEncoderConfiguration(
@@ -57,48 +45,6 @@ class SimpleFeignEncoderConfiguration(
 
     @Bean
     fun simpleErrorDecoder(): ErrorDecoder {
-        return CustomErrorDecoder()
+        return CustomFeignErrorDecoder()
     }
-
-    class CustomErrorDecoder : ErrorDecoder {
-        private val logger = LoggerFactory.getLogger(this.javaClass)
-        override fun decode(methodKey: String?, response: Response?): Exception {
-            val responseMessage: String? = getMessage(response)
-            return when (response?.status()) {
-                400 -> IllegalArgumentException(responseMessage)
-                422 -> BusinessException.of(MooveErrorCode.INVALID_PAYLOAD, responseMessage ?: response.reason())
-                else -> RuntimeException(responseMessage)
-            }
-        }
-
-        private fun getMessage(response: Response?): String {
-            val responseMessage = this.extractMessageFromResponse(response)
-            return  responseMessage ?: "Internal server error"
-        }
-
-        private fun extractMessageFromResponse(response: Response?): String? {
-            var responseAsString: String? = null
-            try {
-                responseAsString = response?.body()?.let {
-                    StreamUtils.copyToString(it.asInputStream(), StandardCharsets.UTF_8)
-                }
-                return responseAsString?.let {
-                    getResponseAsObject(it)
-                }
-            } catch (ex: IOException) {
-                logger.error(ex.message, ex)
-                return responseAsString ?: "Error reading response of request"
-            }
-        }
-        private fun getResponseAsObject(message: String): String {
-            val objectResponse = jacksonObjectMapper().readValue(message, ErrorResponse::class.java)
-            return objectResponse.message.toString()
-        }
-    }
-    @JsonIgnoreProperties(ignoreUnknown = true)
-    data class ErrorResponse(
-        val statusCode: String,
-        val message: Any,
-        val error: String
-    )
 }
